fix: fix card item positioning + Custom card moved to their screen class
