+++ app/src/main/java/ir/jaamebaade/jaamebaade_client/view/MyBookmarkScreen.kt
@@ -38,6 +38,8 @@ import androidx.navigation.NavController
 import ir.jaamebaade.jaamebaade_client.R
 import ir.jaamebaade.jaamebaade_client.constants.AppRoutes
 import ir.jaamebaade.jaamebaade_client.model.BookmarkPoemCategoriesPoetFirstVerse
+import ir.jaamebaade.jaamebaade_client.model.Category
+import ir.jaamebaade.jaamebaade_client.model.Poem
 import ir.jaamebaade.jaamebaade_client.view.components.CardItem
 import ir.jaamebaade.jaamebaade_client.view.components.bookmark.BottomSheetListItem
 import ir.jaamebaade.jaamebaade_client.view.components.toast.ToastType
@@ -95,7 +97,7 @@ fun MyBookmarkScreen(
                 key = { _, item -> item.bookmark.id }) { index, bookmark ->
                 CardItem(
                     modifier = Modifier.animateItem(),
-                    headerText = AnnotatedString(bookmark.poem.title),
+                    headerText = AnnotatedString(createPoemPath(bookmark.categories, bookmark.poem)),
                     bodyText = bookmark.firstVerse.text,
                     imageUrl = bookmark.poet.imageUrl,
                     onClick = { navController.navigate("${AppRoutes.POEM}/${bookmark.poet.id}/${bookmark.poem.id}/-1") },
@@ -154,3 +156,6 @@ fun MyBookmarkScreen(
         }
     }
 }
+
+private fun createPoemPath(categories: List<Category>, poem: Poem) =
+    "${categories.joinToString(" > ") { it.text }} > ${poem.title}"
