+++ app/src/main/java/ir/jaamebaade/jaamebaade_client/view/components/CardItem.kt
@@ -2,16 +2,13 @@ package ir.jaamebaade.jaamebaade_client.view.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Circle
-import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Icon
@@ -23,18 +20,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import ir.jaamebaade.jaamebaade_client.R
-import ir.jaamebaade.jaamebaade_client.model.CommentPoemCategoriesPoet
-import ir.jaamebaade.jaamebaade_client.model.HistoryRecordPathFirstVerse
-import ir.jaamebaade.jaamebaade_client.ui.theme.secondaryS50
-import ir.jaamebaade.jaamebaade_client.utility.convertToJalali
-import ir.jaamebaade.jaamebaade_client.utility.toLocalFormatWithHour
 import ir.jaamebaade.jaamebaade_client.view.components.base.SquareImage
-import java.util.Date
 
 @Composable
 fun CardItem(
@@ -90,7 +79,7 @@ fun CardItem(
                             modifier = Modifier.weight(7f),
                             text = it,
                             style = MaterialTheme.typography.headlineMedium,
-                            maxLines = if (wrapHeader) Int.MAX_VALUE else 1,
+                            maxLines = if (wrapHeader) Int.MAX_VALUE else 3,
                             overflow = if (wrapHeader) TextOverflow.Clip else TextOverflow.Ellipsis
                         )
                     }
@@ -136,32 +125,6 @@ fun CardItem(
 
 
 @Composable
-fun HighlightCardItem(
-    modifier: Modifier = Modifier,
-    headerText: AnnotatedString? = null,
-    bodyText: String,
-    imageUrl: String?,
-    icon: ImageVector? = null,
-    iconDescription: String? = null,
-    onClick: () -> Unit,
-    onIconClick: () -> Unit = {},
-) {
-    CardItem(
-        modifier = modifier,
-        headerText = headerText,
-        bodyText = bodyText,
-        imageUrl = imageUrl,
-        icon = icon,
-        iconDescription = iconDescription,
-        onClick = onClick,
-        onIconClick = onIconClick,
-        wrapHeader = true,
-        wrapBody = true,
-    )
-
-}
-
-@Composable
 fun ComposableCardItem(
     modifier: Modifier = Modifier,
     header: (@Composable () -> Unit)? = null,
@@ -188,6 +151,8 @@ fun ComposableCardItem(
                 .fillMaxWidth()
                 .padding(vertical = 12.dp),
             horizontalArrangement = Arrangement.SpaceBetween,
+            verticalAlignment = Alignment.CenterVertically
+
         ) {
             SquareImage(
                 imageUrl = imageUrl,
@@ -202,19 +167,19 @@ fun ComposableCardItem(
             ) {
                 Row(
                     modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(end = 24.dp),
+                        .fillMaxWidth(),
+                       // .padding(end = 12.dp),
                     horizontalArrangement = Arrangement.SpaceBetween,
                     verticalAlignment = Alignment.CenterVertically
                 ) {
-                    Row(
-                        modifier = Modifier.fillMaxWidth().weight(7f)
+                    Box(
+                        modifier = Modifier.weight(7f)
                     ) {
-                        Column() {
+                        //Column() {
                             header?.let { it() }
-                            body()
 
-                        }
+
+                        //}
                     }
                     if (icon != null) {
                         IconButton(
@@ -231,117 +196,9 @@ fun ComposableCardItem(
                     }
 
                 }
-
+                body()
                 footer?.let { it() }
             }
         }
     }
-}
-
-@Composable
-fun MyNoteCardItem(
-    modifier: Modifier = Modifier,
-    note: CommentPoemCategoriesPoet,
-    onClick: () -> Unit = {},
-    onIconClick: () -> Unit = {},
-) {
-    ComposableCardItem(
-        modifier = modifier,
-        imageUrl = note.path.poet.imageUrl,
-        header = {
-            Text(
-                text = note.path.poem.title,
-                style = MaterialTheme.typography.headlineMedium,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-        },
-        body = {
-            Text(
-                text = note.comment.text.trim(),
-                style = MaterialTheme.typography.headlineSmall,
-                color = MaterialTheme.colorScheme.outlineVariant,
-            )
-        },
-        footer =
-            {
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Icon(
-                        imageVector = Icons.Default.Circle,
-                        tint = MaterialTheme.colorScheme.secondaryS50,
-                        contentDescription = null,
-                        modifier = Modifier.size(8.dp),
-                    )
-                    Spacer(modifier = Modifier.width(8.dp))
-                    Text(
-                        text = Date(note.comment.createdAt).convertToJalali()
-                            .toLocalFormatWithHour(),
-                        style = MaterialTheme.typography.headlineSmall,
-                        color = MaterialTheme.colorScheme.outlineVariant,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis,
-                    )
-                }
-
-            },
-        icon = Icons.Filled.MoreVert,
-        iconDescription = stringResource(R.string.MORE),
-        onClick = onClick,
-        onIconClick = onIconClick
-    )
-}
-
-
-@Composable
-fun MyHistoryCardItem(
-    modifier: Modifier = Modifier,
-    historyRecord: HistoryRecordPathFirstVerse,
-    onClick: () -> Unit = {},
-) {
-    ComposableCardItem(
-        modifier = modifier,
-        imageUrl = historyRecord.path.poet.imageUrl,
-        header = {
-            Text(
-                text = historyRecord.path.poem.title,
-                style = MaterialTheme.typography.headlineMedium,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-        },
-        body = {
-            Text(
-                text = historyRecord.firstVerse.text,
-                style = MaterialTheme.typography.headlineSmall,
-                color = MaterialTheme.colorScheme.outlineVariant,
-                maxLines = 2,
-                overflow = TextOverflow.Ellipsis,
-            )
-        },
-        footer =
-            {
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Icon(
-                        imageVector = Icons.Default.Circle,
-                        tint = MaterialTheme.colorScheme.secondaryS50,
-                        contentDescription = null,
-                        modifier = Modifier.size(8.dp),
-                    )
-                    Spacer(modifier = Modifier.width(8.dp))
-                    Text(
-                        text = Date(historyRecord.timestamp).convertToJalali()
-                            .toLocalFormatWithHour(),
-                        style = MaterialTheme.typography.headlineSmall,
-                        color = MaterialTheme.colorScheme.outlineVariant,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis,
-                    )
-                }
-
-            },
-        icon = null,
-        iconDescription = null,
-        onClick = onClick,
-        containerColor = MaterialTheme.colorScheme.background
-    )
 }
